[FILE] fonts: DejaVu Sans Mono for powerline
Normally I am working on a Linux machine using vim and tmux with the powerline statusline plugin.
However, sometimes I am forced to work on a windows machine. Whenever possible I open a ssh connection with PuTTY to my Linux machine to work remote on it.
It turned out that the fonts of powerline were not displayed correctly.
To fix this, please install these fonts.